TranskribusAPI_Webapp/requirements.txt
Streamlit
streamlit_option_menu
pandas
pillow
streamlit_extras